Just figured out GitHub
Couple days of work: never put it up so here's a bunch in the order they
appear in the code.

69 - Added error handing for the class info coming up wrong
100 - You can set the program to drop a CRN when a new one is being
added.  So say you need a class you already have but want a different
time, it will drop it for you.
130 - Now shows the name you enter at the start of the program, helpful
if you have a couple different people running at the same time.
156 - Keeps track of what classes have been added and shows them
181 - The code for handling a CRN to be removed
196 - If all the classes are added the program stops
205 - If you enter the wrong login it will loop back and not end the
program